Adds option to allow connections in test.
This is useful in CI environments where you want to hit a real service.

When not set, it will keep existing behavior, which is to block all
requests in the test env.

When set to true, the environment variable will determine whether a
service is mocked or not.

When set to false, all connections will be blocked, even if enabled via
environment variable.

It defaults to false in test and true in other environments.